requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.4
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
arrow==1.2.3
asteroid-filterbanks==0.4.0
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
backports.cached-property==1.0.2
beautifulsoup4==4.12.2
blessed==1.20.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmaes==0.9.1
colorama==0.4.6
colorlog==6.7.0
commonmark==0.9.1
contourpy==1.0.7
croniter==1.3.14
cycler==0.11.0
dateutils==0.6.12
decorator==5.1.1
deepdiff==6.3.0
docopt==0.6.2
einops==0.6.1
fastapi==0.88.0
ffmpeg-python==0.2.0
ffprobe==0.5
filelock==3.12.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
future==0.18.3
google-auth==2.17.3
google-auth-oauthlib==1.0.0
greenlet==2.0.2
grpcio==1.54.0
h11==0.14.0
hmmlearn @ git+https://github.com/hmmlearn/hmmlearn.git@822894e24e8632cfa3f10304240d3f5049516580
huggingface-hub==0.14.1
HyperPyYAML==1.2.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
inquirer==3.1.3
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
julius==0.2.7
kiwisolver==1.4.4
lazy_loader==0.2
librosa==0.9.2
light-the-torch==0.7.3
lightning==2.0.2
lightning-cloud==0.5.34
lightning-fabric==2.0.2
lightning-utilities==0.8.0
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
networkx==2.8.8
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
openai-whisper @ git+https://github.com/openai/whisper.git@c09a7ae299c4c34c5839a76380ae407e7d785914
optuna==3.1.1
ordered-set==4.1.0
packaging==23.1
pandas==2.0.1
pescador==2.1.0
Pillow==9.5.0
pooch==1.6.0
primePy==1.3
protobuf==3.20.1
psutil==5.9.5
pyannote.audio @ https://github.com/pyannote/pyannote-audio/archive/refs/heads/develop.zip#sha256=27e72e87c28d555cfe169dd932ffc454943e394e8801e75f91d7cea851b5f160
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==2.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
pyDeprecate==0.3.2
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==1.6.5
pytorch-metric-learning==2.1.0
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
readchar==4.0.5
regex==2023.3.23
requests==2.29.0
requests-oauthlib==1.3.1
resampy==0.4.2
rich==12.6.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.2
scipy==1.10.1
semver==3.0.0
sentencepiece==0.1.98
shellingham==1.5.0.post1
simplejson==3.19.1
singledispatchmethod==1.0
six==1.16.0
sniffio==1.3.0
sortedcollections==2.1.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.5
speechbrain==0.5.14
SQLAlchemy==2.0.11
starlette==0.22.0
starsessions==1.3.0
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
threadpoolctl==3.1.0
tiktoken==0.3.1
torch==2.0.0
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchaudio==2.0.1
torchmetrics==0.11.4
torchvision==0.15.1
tqdm==4.65.0
traitlets==5.9.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.21.1
wcwidth==0.2.6
websocket-client==1.5.1
websockets==11.0.2
Werkzeug==2.3.1
whisper==1.1.10
yarl==1.9.2
zipp==3.15.0
